# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Production-grade Go backend for Hackathon Intelligence.",
        "title": "Hackathon Intelligence Platform API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "https://hackbuddy-im66.onrender.com",
    "basePath": "/api/v1",
    "paths": {
        "/admin/analyses": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all analyses. Admin only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List all analyses (admin)",
                "responses": {
                    "200": {
                        "description": "Analyses array in data",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "Admin required",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/admin/model": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns current active model (from override or config) and list of available model IDs.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get active LLM model and available models (admin)",
                "responses": {
                    "200": {
                        "description": "active, available in data",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "Admin required",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Override the active model for all LLM calls. Use a value from GET /admin/model available list.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Set active LLM model (admin)",
                "parameters": [
                    {
                        "description": "Model ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domains_admin.SetModelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "active model in data",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid body",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "Admin required",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/admin/role/{user_id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Set user role to user or admin. Admin only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update user role (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Role",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domains_admin.UpdateRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role updated",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid role",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "Admin required",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/admin/sessions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all sessions. Admin only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List all sessions (admin)",
                "responses": {
                    "200": {
                        "description": "Sessions array in data",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "Admin required",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/admin/usage": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns token usage rows. Optional query: user_id, session_id to filter.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List token usage (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by user ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by session ID",
                        "name": "session_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token usage array in data",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "Admin required",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/admin/usage/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns aggregated token usage for gauging and manual model switching.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Token usage summary by user/session/model (admin)",
                "responses": {
                    "200": {
                        "description": "Usage summary array in data",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "Admin required",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/admin/user/{user_id}/delete": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Invalidates the user; they can be permanently removed after 90 days. Admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Soft delete a user (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User soft-deleted",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "Admin required",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/admin/user/{user_id}/hard-delete": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes the user and all associated data. Admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Permanently delete a user (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "Admin required",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all users. Admin only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List all users (admin)",
                "responses": {
                    "200": {
                        "description": "Users array in data",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "Admin required",
                        "schema": {
                            "$ref": "#/definitions/hackbuddy-backend_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domains_admin.SetModelRequest": {
            "type": "object",
            "required": [
                "model"
            ],
            "properties": {
                "model": {
                    "type": "string"
                }
            }
        },
        "domains_admin.UpdateRoleRequest": {
            "type": "object",
            "required": [
                "role"
            ],
            "properties": {
                "role": {
                    "type": "string",
                    "enum": [
                        "user",
                        "admin"
                    ]
                }
            }
        },
...
        "hackbuddy-backend_pkg_response.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "{} | [] | null"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "description": "\"success\" | \"error\"",
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}